Refactor Piece
Set the previous Piece.rules method to be a field now, and add babel
transform-class-properties to be able to set it without a constructor,
this will improve the creation of new pieces.
Also, set the Piece.possiblePositions method to be an internal method,
that will be used by Piece.possiblePositionsInNMoves.